+++ CHANGES.md
@@ -15,10 +15,12 @@
     - @miosakuma
 - [ADD] 転送フィルター機能を追加する
     - @szktty
-- [ADD]
-  - scalability mode に対応する
+- [ADD] scalability mode に対応する
   - VP9 / AV1 のサイマルキャストに対応可能になる
   - @szktty
+- [ADD] 映像コーデックパラメータを追加する
+    - `SoraMediaOption` に `videoVp9Params`, `videoAv1Params`, `videoH264Params` を追加する
+    - @miosakuma
 
 ## 2023.1.0